requirements.txt
numpy~=1.24.4
scipy~=1.10.1
joblib~=1.4.0
gymnasium~=0.28.1
matplotlib~=3.7.5
aiohttp~=3.9.5
ray~=2.10.0
scikit-learn~=1.3.2
tensorflow == 2.12
tensorflow-estimator == 2.12.0
tensorflow-io-gcs-filesystem == 0.26.0
tensorflow-probability == 0.20.1
paho-mqtt >= 1.6.1, < 1.7
pyarrow~=15.0.2
typer==0.12.3
pandas==2.0.3
bootstrapped==0.0.2
